ALTANMYA-PricingBasedOnBom/requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
apturl==0.5.2
async-timeout==3.0.1
attrs==21.4.0
Babel==2.6.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.10
cached-property==1.5.2
cachetools==5.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbfread==2.0.7
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
ebaysdk==2.1.5
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
feedparser==5.2.1
firebase-admin==5.2.0
freezegun==0.3.15
future==0.18.2
gevent==20.9.0
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
greenlet==0.4.17
grpcio==1.44.0
grpcio-status==1.44.0
html2text==2020.1.16
html5lib==1.0.1
httplib2==0.20.4
idna==2.8
isodate==0.6.1
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libsass==0.18.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mock==3.0.5
monotonic==1.5
msgpack==1.0.3
multidict==4.7.3
netifaces==0.10.4
num2words==0.5.6
numpy==1.22.3
oauthlib==3.1.0
odoo==15.0+e.20220314
ofxparse==0.19
olefile==0.46
openpyxl==3.0.10
pandas==1.4.3
paramiko==2.6.0
passlib==1.7.2
pbr==5.4.5
pdfminer.six==20211012
pexpect==4.6.0
Pillow==8.1.2
polib==1.1.0
proto-plus==1.20.3
protobuf==3.6.1
psutil==5.6.6
psycopg2==2.8.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pydot==1.4.1
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyRFC3339==1.1
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-ldap==3.2.0
python-stdnum==1.13
pytz==2022.1
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.59
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.8
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.1
ssh-import-id==5.10
systemd-python==234
systemd-socketserver==1.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vobject==0.9.6.1
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==0.16.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
yarl==1.4.2
zeep==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
